Geometry property results have been added

+++ StructureHelper/Windows/ViewModels/Materials/MaterialsViewModel.cs
@@ -6,16 +6,40 @@ using StructureHelperCommon.Infrastructures.Settings;
 using StructureHelperCommon.Infrastructures.Strings;
 using StructureHelperCommon.Models.Forces;
 using StructureHelperLogics.Models.Materials;
+using StructureHelperLogics.NdmCalculations.Primitives;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.TextBox;
+using System.Windows.Forms;
+using StructureHelperLogics.Models.CrossSections;
+using StructureHelper.Infrastructure;
+using System.Windows.Input;
 
 namespace StructureHelper.Windows.ViewModels.Materials
 {
-    internal class MaterialsViewModel : CRUDViewModelBase<IHeadMaterial>
+    public class MaterialsViewModel : CRUDViewModelBase<IHeadMaterial>
     {
+        ICrossSectionRepository repository;
+        private ICommand editMaterialsCommand;
+
+        public ICommand EditMaterialsCommand
+        {
+            get
+            {
+                return editMaterialsCommand ??
+                    (
+                    editMaterialsCommand = new RelayCommand(o => EditHeadMaterials())
+                    );
+            }
+            
+        }
+        public MaterialsViewModel(ICrossSectionRepository repository) : base(repository.HeadMaterials)
+        {
+            this.repository = repository;
+        }
         public override void AddMethod(object parameter)
         {
             CheckParameters(parameter);
@@ -28,9 +52,19 @@ namespace StructureHelper.Windows.ViewModels.Materials
         }
         public override void DeleteMethod(object parameter)
         {
-#error
-            //to do delete method
-            base.DeleteMethod(parameter);
+            var primitives = repository.Primitives;
+            var primitivesWithMaterial = primitives.Where(x => x.HeadMaterial == SelectedItem);
+            int primitivesCount = primitivesWithMaterial.Count();
+            if (primitivesCount > 0)
+            {
+                MessageBox.Show("Some primitives reference to this material", "Material can not be deleted", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+            var dialogResult = MessageBox.Show("Delete material?", "Please, confirm deleting", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
+            {
+                base.DeleteMethod(parameter);
+            }
         }
         public override void EditMethod(object parameter)
         {
@@ -38,7 +72,6 @@ namespace StructureHelper.Windows.ViewModels.Materials
             wnd.ShowDialog();
             base.EditMethod(parameter);
         }
-
         private void AddElastic()
         {
             var material = HeadMaterialFactory.GetHeadMaterial(HeadmaterialType.Elastic200, ProgramSetting.CodeType);
@@ -61,6 +94,12 @@ namespace StructureHelper.Windows.ViewModels.Materials
         {
             if (parameter is null) { throw new StructureHelperException(ErrorStrings.ParameterIsNull); }
             if (parameter is not MaterialType) { throw new StructureHelperException(ErrorStrings.ObjectTypeIsUnknown + $". Expected: {typeof(MaterialType)} . Actual type: {nameof(parameter)}"); }
+        }
+        private void EditHeadMaterials()
+        {
+            var wnd = new HeadMaterialsView(repository);
+            wnd.ShowDialog();
+            Refresh();
         }
     }
 }
